Add cohorts metadata support for Af1
