+++ malariagen_data/anopheles.py
@@ -36,6 +36,8 @@ from .util import (
     dask_compress_dataset,
     init_filesystem,
     init_zarr_store,
+    jackknife_ci,
+    jitter,
     locate_region,
     read_gff3,
     resolve_region,
@@ -283,7 +285,6 @@ class AnophelesDataResource(ABC):
     @abstractmethod
     def _sample_metadata(self, sample_set):
         # Not all children have species calls or cohorts data.
-        # TODO: maybe _read_species_calls and _read_cohort_metadata return empty.
         raise NotImplementedError("Must override _sample_metadata")
 
     @abstractmethod
@@ -312,7 +313,7 @@ class AnophelesDataResource(ABC):
         cohort_size,
         random_seed,
     ):
-        # Ag3 has cohorts and species. Subclasses have different cache names.
+        # Subclasses have different cache names.
         raise NotImplementedError("Must override snp_allele_counts")
 
     @abstractmethod
@@ -331,7 +332,23 @@ class AnophelesDataResource(ABC):
         ci_method,
     ):
         # Subclasses have different variant filter pass fields.
+        # TODO: this is only a small difference, which might be possible to abstract.
         raise NotImplementedError("Must override snp_allele_frequencies_advanced")
+
+    @abstractmethod
+    def fst_gwss(
+        self,
+        contig,
+        window_size,
+        cohort1_query,
+        cohort2_query,
+        sample_sets,
+        site_mask,
+        cohort_size,
+        random_seed,
+    ):
+        # Subclasses have different cache names.
+        raise NotImplementedError("Must override fst_gwss")
 
     # Start of @staticmethod @abstractmethod
 
@@ -2149,6 +2166,7 @@ class AnophelesDataResource(ABC):
         )
 
         debug("perform allele count")
+        # FIXME: Parameter 'cohort_size', 'random_seed', 'site_class' unfilled
         ac = self.snp_allele_counts(
             region=region,
             sample_sets=sample_sets,
@@ -3548,6 +3566,7 @@ class AnophelesDataResource(ABC):
             raise ValueError("Region is too large, please provide a smaller region.")
 
         debug("compute allele counts")
+        # FIXME: Parameters 'random_seed', 'site_class' unfilled
         ac = allel.AlleleCountsArray(
             self.snp_allele_counts(
                 region=region,
@@ -4195,3 +4214,1634 @@ class AnophelesDataResource(ABC):
         ds_aa_frq.attrs["title"] = title
 
         return ds_aa_frq
+
+    def _block_jackknife_cohort_diversity_stats(
+        self, *, cohort_label, ac, n_jack, confidence_level
+    ):
+        debug = self._log.debug
+
+        debug("set up for diversity calculations")
+        n_sites = ac.shape[0]
+        ac = allel.AlleleCountsArray(ac)
+        n = ac.sum(axis=1).max()  # number of chromosomes sampled
+        n_sites = min(n_sites, ac.shape[0])  # number of sites
+        block_length = n_sites // n_jack  # number of sites in each block
+        n_sites_j = n_sites - block_length  # number of sites in each jackknife resample
+
+        debug("compute scaling constants")
+        a1 = np.sum(1 / np.arange(1, n))
+        a2 = np.sum(1 / (np.arange(1, n) ** 2))
+        b1 = (n + 1) / (3 * (n - 1))
+        b2 = 2 * (n**2 + n + 3) / (9 * n * (n - 1))
+        c1 = b1 - (1 / a1)
+        c2 = b2 - ((n + 2) / (a1 * n)) + (a2 / (a1**2))
+        e1 = c1 / a1
+        e2 = c2 / (a1**2 + a2)
+
+        debug(
+            "compute some intermediates ahead of time, to minimise computation during jackknife resampling"
+        )
+        mpd_data = allel.mean_pairwise_difference(ac, fill=0)
+        # N.B., here we compute the number of segregating sites as the number
+        # of alleles minus 1. This follows the sgkit and tskit implementations,
+        # and is different from scikit-allel.
+        seg_data = ac.allelism() - 1
+
+        debug("compute estimates from all data")
+        # FIXME: variable in function should be lowercase
+        theta_pi_abs_data = np.sum(mpd_data)
+        theta_pi_data = theta_pi_abs_data / n_sites
+        S_data = np.sum(seg_data)
+        theta_w_abs_data = S_data / a1
+        theta_w_data = theta_w_abs_data / n_sites
+        d_data = theta_pi_abs_data - theta_w_abs_data
+        d_stdev_data = np.sqrt((e1 * S_data) + (e2 * S_data * (S_data - 1)))
+        tajima_d_data = d_data / d_stdev_data
+
+        debug("set up for jackknife resampling")
+        jack_theta_pi = []
+        jack_theta_w = []
+        jack_tajima_d = []
+
+        debug("begin jackknife resampling")
+        for i in range(n_jack):
+
+            # locate block to delete
+            block_start = i * block_length
+            block_stop = block_start + block_length
+            loc_j = np.ones(n_sites, dtype=bool)
+            loc_j[block_start:block_stop] = False
+            assert np.count_nonzero(loc_j) == n_sites_j
+
+            # resample data and compute statistics
+
+            # theta_pi
+            mpd_j = mpd_data[loc_j]
+            theta_pi_abs_j = np.sum(mpd_j)
+            theta_pi_j = theta_pi_abs_j / n_sites_j
+            jack_theta_pi.append(theta_pi_j)
+
+            # theta_w
+            # FIXME: variable in function should be lowercase
+            seg_j = seg_data[loc_j]
+            S_j = np.sum(seg_j)
+            theta_w_abs_j = S_j / a1
+            theta_w_j = theta_w_abs_j / n_sites_j
+            jack_theta_w.append(theta_w_j)
+
+            # tajima_d
+            d_j = theta_pi_abs_j - theta_w_abs_j
+            d_stdev_j = np.sqrt((e1 * S_j) + (e2 * S_j * (S_j - 1)))
+            tajima_d_j = d_j / d_stdev_j
+            jack_tajima_d.append(tajima_d_j)
+
+        # calculate jackknife stats
+        (
+            theta_pi_estimate,
+            theta_pi_bias,
+            theta_pi_std_err,
+            theta_pi_ci_err,
+            theta_pi_ci_low,
+            theta_pi_ci_upp,
+        ) = jackknife_ci(
+            stat_data=theta_pi_data,
+            jack_stat=jack_theta_pi,
+            confidence_level=confidence_level,
+        )
+        (
+            theta_w_estimate,
+            theta_w_bias,
+            theta_w_std_err,
+            theta_w_ci_err,
+            theta_w_ci_low,
+            theta_w_ci_upp,
+        ) = jackknife_ci(
+            stat_data=theta_w_data,
+            jack_stat=jack_theta_w,
+            confidence_level=confidence_level,
+        )
+        (
+            tajima_d_estimate,
+            tajima_d_bias,
+            tajima_d_std_err,
+            tajima_d_ci_err,
+            tajima_d_ci_low,
+            tajima_d_ci_upp,
+        ) = jackknife_ci(
+            stat_data=tajima_d_data,
+            jack_stat=jack_tajima_d,
+            confidence_level=confidence_level,
+        )
+
+        return dict(
+            cohort=cohort_label,
+            theta_pi=theta_pi_data,
+            theta_pi_estimate=theta_pi_estimate,
+            theta_pi_bias=theta_pi_bias,
+            theta_pi_std_err=theta_pi_std_err,
+            theta_pi_ci_err=theta_pi_ci_err,
+            theta_pi_ci_low=theta_pi_ci_low,
+            theta_pi_ci_upp=theta_pi_ci_upp,
+            theta_w=theta_w_data,
+            theta_w_estimate=theta_w_estimate,
+            theta_w_bias=theta_w_bias,
+            theta_w_std_err=theta_w_std_err,
+            theta_w_ci_err=theta_w_ci_err,
+            theta_w_ci_low=theta_w_ci_low,
+            theta_w_ci_upp=theta_w_ci_upp,
+            tajima_d=tajima_d_data,
+            tajima_d_estimate=tajima_d_estimate,
+            tajima_d_bias=tajima_d_bias,
+            tajima_d_std_err=tajima_d_std_err,
+            tajima_d_ci_err=tajima_d_ci_err,
+            tajima_d_ci_low=tajima_d_ci_low,
+            tajima_d_ci_upp=tajima_d_ci_upp,
+        )
+
+    def cohort_diversity_stats(
+        self,
+        cohort,
+        cohort_size,
+        region,
+        site_mask,
+        site_class=None,
+        sample_sets=None,
+        random_seed=42,
+        n_jack=200,
+        confidence_level=0.95,
+    ):
+        """Compute genetic diversity summary statistics for a cohort of
+        individuals.
+
+        Parameters
+        ----------
+        cohort : str or (str, str)
+            Either a string giving one of the predefined cohort labels, or a
+            pair of strings giving a custom cohort label and a sample query to
+            select samples in the cohort.
+        cohort_size : int
+            Number of individuals to use for computation of summary statistics.
+            If the cohort is larger than this size, it will be randomly
+            down-sampled.
+        region : str
+            Chromosome arm, gene name or
+            genomic region defined with coordinates (e.g.,
+            "2L:44989425-44998059").
+        site_mask : str
+            Site filters mask to apply.
+        site_class : str, optional
+            Select sites belonging to one of the following classes: CDS_DEG_4,
+            (4-fold degenerate coding sites), CDS_DEG_2_SIMPLE (2-fold simple
+            degenerate coding sites), CDS_DEG_0 (non-degenerate coding sites),
+            INTRON_SHORT (introns shorter than 100 bp), INTRON_LONG (introns
+            longer than 200 bp), INTRON_SPLICE_5PRIME (intron within 2 bp of
+            5' splice site), INTRON_SPLICE_3PRIME (intron within 2 bp of 3'
+            splice site), UTR_5PRIME (5' untranslated region), UTR_3PRIME (3'
+            untranslated region), INTERGENIC (intergenic, more than 10 kbp from
+            a gene).
+        sample_sets : str or list of str, optional
+            Can be a sample set identifier or a list of
+            sample set identifiers  or a
+            release identifier or a list of release identifiers.
+        random_seed : int, optional
+            Seed for random number generator.
+        n_jack : int, optional
+            Number of blocks to divide the data into for the block jackknife
+            estimation of confidence intervals. N.B., larger is not necessarily
+            better.
+        confidence_level : float, optional
+            Confidence level to use for confidence interval calculation. 0.95
+            means 95% confidence interval.
+
+        Returns
+        -------
+        stats : pandas.Series
+            A series with summary statistics and their confidence intervals.
+
+        """
+
+        debug = self._log.debug
+
+        debug("process cohort parameter")
+        cohort_query = None
+        if isinstance(cohort, str):
+            # assume it is one of the predefined cohorts
+            cohort_label = cohort
+            df_samples = self.sample_metadata(sample_sets=sample_sets)
+            cohort_cols = [c for c in df_samples.columns if c.startswith("cohort_")]
+            for c in cohort_cols:
+                if cohort in set(df_samples[c]):
+                    cohort_query = f"{c} == '{cohort}'"
+                    break
+            if cohort_query is None:
+                raise ValueError(f"unknown cohort: {cohort}")
+
+        elif isinstance(cohort, (list, tuple)) and len(cohort) == 2:
+            cohort_label, cohort_query = cohort
+
+        else:
+            raise TypeError(r"invalid cohort parameter: {cohort!r}")
+
+        debug("access allele counts")
+        ac = self.snp_allele_counts(
+            region=region,
+            site_mask=site_mask,
+            site_class=site_class,
+            sample_query=cohort_query,
+            sample_sets=sample_sets,
+            cohort_size=cohort_size,
+            random_seed=random_seed,
+        )
+
+        debug("compute diversity stats")
+        stats = self._block_jackknife_cohort_diversity_stats(
+            cohort_label=cohort_label,
+            ac=ac,
+            n_jack=n_jack,
+            confidence_level=confidence_level,
+        )
+
+        debug("compute some extra cohort variables")
+        df_samples = self.sample_metadata(
+            sample_sets=sample_sets, sample_query=cohort_query
+        )
+        extra_fields = [
+            ("taxon", "unique"),
+            ("year", "unique"),
+            ("month", "unique"),
+            ("country", "unique"),
+            ("admin1_iso", "unique"),
+            ("admin1_name", "unique"),
+            ("admin2_name", "unique"),
+            ("longitude", "mean"),
+            ("latitude", "mean"),
+        ]
+        for field, agg in extra_fields:
+            if agg == "unique":
+                vals = df_samples[field].sort_values().unique()
+                if len(vals) == 0:
+                    val = np.nan
+                elif len(vals) == 1:
+                    val = vals[0]
+                else:
+                    val = vals.tolist()
+            elif agg == "mean":
+                vals = df_samples[field]
+                if len(vals) == 0:
+                    val = np.nan
+                else:
+                    val = np.mean(vals)
+            else:
+                val = np.nan
+            stats[field] = val
+
+        return pd.Series(stats)
+
+    # TODO: compare with cohort_diversity_stats()
+    def diversity_stats(
+        self,
+        cohorts,
+        cohort_size,
+        region,
+        site_mask,
+        site_class=None,
+        sample_query=None,
+        sample_sets=None,
+        random_seed=42,
+        n_jack=200,
+        confidence_level=0.95,
+    ):
+        """Compute genetic diversity summary statistics for multiple cohorts.
+
+        Parameters
+        ----------
+        cohorts : str or dict
+            Either a string giving one of the predefined cohort columns, or a
+            dictionary mapping cohort labels to sample queries.
+        cohort_size : int
+            Number of individuals to use for computation of summary statistics.
+            If the cohort is larger than this size, it will be randomly
+            down-sampled.
+        region : str
+            Chromosome arm (e.g., "2L"), gene name (e.g., "AGAP007280") or
+            genomic region defined with coordinates (e.g.,
+            "2L:44989425-44998059").
+        site_mask : {"gamb_colu_arab", "gamb_colu", "arab"}
+            Site filters mask to apply.
+        site_class : str, optional
+            Select sites belonging to one of the following classes: CDS_DEG_4,
+            (4-fold degenerate coding sites), CDS_DEG_2_SIMPLE (2-fold simple
+            degenerate coding sites), CDS_DEG_0 (non-degenerate coding sites),
+            INTRON_SHORT (introns shorter than 100 bp), INTRON_LONG (introns
+            longer than 200 bp), INTRON_SPLICE_5PRIME (intron within 2 bp of
+            5' splice site), INTRON_SPLICE_3PRIME (intron within 2 bp of 3'
+            splice site), UTR_5PRIME (5' untranslated region), UTR_3PRIME (3'
+            untranslated region), INTERGENIC (intergenic, more than 10 kbp from
+            a gene).
+        sample_query : str, optional
+            A pandas query string which will be evaluated against the sample
+            metadata e.g., "taxon == 'coluzzii' and country == 'Burkina Faso'".
+        sample_sets : str or list of str, optional
+            Can be a sample set identifier (e.g., "AG1000G-AO") or a list of
+            sample set identifiers (e.g., ["AG1000G-BF-A", "AG1000G-BF-B"]) or a
+            release identifier (e.g., "3.0") or a list of release identifiers.
+        random_seed : int, optional
+            Seed for random number generator.
+        n_jack : int, optional
+            Number of blocks to divide the data into for the block jackknife
+            estimation of confidence intervals. N.B., larger is not necessarily
+            better.
+        confidence_level : float, optional
+            Confidence level to use for confidence interval calculation. 0.95
+            means 95% confidence interval.
+
+        Returns
+        -------
+        df_stats : pandas.DataFrame
+            A DataFrame where each row provides summary statistics and their
+            confidence intervals for a single cohort.
+
+        """
+        debug = self._log.debug
+        info = self._log.info
+
+        debug("set up cohorts")
+        if isinstance(cohorts, dict):
+            # user has supplied a custom dictionary mapping cohort identifiers
+            # to pandas queries
+            cohort_queries = cohorts
+
+        elif isinstance(cohorts, str):
+            # user has supplied one of the predefined cohort sets
+
+            df_samples = self.sample_metadata(
+                sample_sets=sample_sets, sample_query=sample_query
+            )
+
+            # determine column in dataframe - allow abbreviation
+            if cohorts.startswith("cohort_"):
+                cohorts_col = cohorts
+            else:
+                cohorts_col = "cohort_" + cohorts
+            if cohorts_col not in df_samples.columns:
+                raise ValueError(f"{cohorts_col!r} is not a known cohort set")
+
+            # find cohort labels and build queries dictionary
+            cohort_labels = sorted(df_samples[cohorts_col].dropna().unique())
+            cohort_queries = {coh: f"{cohorts_col} == '{coh}'" for coh in cohort_labels}
+
+        else:
+            raise TypeError("cohorts parameter should be dict or str")
+
+        debug("handle sample_query parameter")
+        if sample_query is not None:
+            cohort_queries = {
+                cohort_label: f"({cohort_query}) and ({sample_query})"
+                for cohort_label, cohort_query in cohort_queries.items()
+            }
+
+        debug("check cohort sizes, drop any cohorts which are too small")
+        cohort_queries_checked = dict()
+        for cohort_label, cohort_query in cohort_queries.items():
+            df_cohort_samples = self.sample_metadata(
+                sample_sets=sample_sets, sample_query=cohort_query
+            )
+            n_samples = len(df_cohort_samples)
+            if n_samples < cohort_size:
+                info(
+                    f"cohort ({cohort_label}) has insufficient samples ({n_samples}) for requested cohort size ({cohort_size}), dropping"  # noqa
+                )  # noqa
+            else:
+                cohort_queries_checked[cohort_label] = cohort_query
+
+        debug("compute diversity stats for cohorts")
+        all_stats = []
+        for cohort_label, cohort_query in cohort_queries_checked.items():
+            stats = self.cohort_diversity_stats(
+                cohort=(cohort_label, cohort_query),
+                cohort_size=cohort_size,
+                region=region,
+                site_mask=site_mask,
+                site_class=site_class,
+                sample_sets=sample_sets,
+                random_seed=random_seed,
+                n_jack=n_jack,
+                confidence_level=confidence_level,
+            )
+            all_stats.append(stats)
+        df_stats = pd.DataFrame(all_stats)
+
+        return df_stats
+
+    def count_samples(
+        self,
+        sample_sets=None,
+        sample_query=None,
+        index=(
+            "country",
+            "admin1_iso",
+            "admin1_name",
+            "admin2_name",
+            "year",
+        ),
+        columns="taxon",
+    ):
+        """Create a pivot table showing numbers of samples available by space,
+        time and taxon.
+
+        Parameters
+        ----------
+        sample_sets : str or list of str, optional
+            Can be a sample set identifier (e.g., "AG1000G-AO") or a list of
+            sample set identifiers (e.g., ["AG1000G-BF-A", "AG1000G-BF-B"]) or a
+            release identifier (e.g., "3.0") or a list of release identifiers.
+        sample_query : str, optional
+            A pandas query string which will be evaluated against the sample
+            metadata e.g., "taxon == 'coluzzii' and country == 'Burkina Faso'".
+        index : str or tuple of str
+            Sample metadata columns to use for the index.
+        columns : str or tuple of str
+            Sample metadata columns to use for the columns.
+
+        Returns
+        -------
+        df : pandas.DataFrame
+            Pivot table of sample counts.
+
+        """
+        debug = self._log.debug
+
+        debug("load sample metadata")
+        df_samples = self.sample_metadata(
+            sample_sets=sample_sets, sample_query=sample_query
+        )
+
+        debug("create pivot table")
+        df_pivot = df_samples.pivot_table(
+            index=index,
+            columns=columns,
+            values="sample_id",
+            aggfunc="count",
+            fill_value=0,
+        )
+
+        return df_pivot
+
+    def _fst_gwss(
+        self,
+        contig,
+        window_size,
+        sample_sets,
+        cohort1_query,
+        cohort2_query,
+        site_mask,
+        cohort_size,
+        random_seed,
+    ):
+
+        ds_snps1 = self.snp_calls(
+            region=contig,
+            sample_query=cohort1_query,
+            sample_sets=sample_sets,
+            site_mask=site_mask,
+            cohort_size=cohort_size,
+            random_seed=random_seed,
+        )
+
+        ds_snps2 = self.snp_calls(
+            region=contig,
+            sample_query=cohort2_query,
+            sample_sets=sample_sets,
+            site_mask=site_mask,
+            cohort_size=cohort_size,
+            random_seed=random_seed,
+        )
+
+        gt1 = allel.GenotypeDaskArray(ds_snps1["call_genotype"].data)
+        with self._dask_progress(desc="Compute allele counts for cohort 1"):
+            ac1 = gt1.count_alleles(max_allele=3).compute()
+
+        gt2 = allel.GenotypeDaskArray(ds_snps2["call_genotype"].data)
+        with self._dask_progress(desc="Compute allele counts for cohort 2"):
+            ac2 = gt2.count_alleles(max_allele=3).compute()
+
+        pos = ds_snps1["variant_position"].values
+
+        fst = allel.moving_hudson_fst(ac1, ac2, size=window_size)
+        x = allel.moving_statistic(pos, statistic=np.mean, size=window_size)
+
+        results = dict(x=x, fst=fst)
+
+        return results
+
+    # Start of plot functions
+
+    def plot_frequencies_heatmap(
+        self,
+        df,
+        index="label",
+        max_len=100,
+        x_label="Cohorts",
+        y_label="Variants",
+        colorbar=True,
+        col_width=40,
+        width=None,
+        row_height=20,
+        height=None,
+        text_auto=".0%",
+        aspect="auto",
+        color_continuous_scale="Reds",
+        title=True,
+        **kwargs,
+    ):
+        """Plot a heatmap from a pandas DataFrame of frequencies, e.g., output
+        from `snp_allele_frequencies()` or `gene_cnv_frequencies()`.
+        It's recommended to filter the input DataFrame to just rows of interest,
+        i.e., fewer rows than `max_len`.
+
+        Parameters
+        ----------
+        df : pandas DataFrame
+           A DataFrame of frequencies, e.g., output from
+           `snp_allele_frequencies()` or `gene_cnv_frequencies()`.
+        index : str or list of str
+            One or more column headers that are present in the input dataframe.
+            This becomes the heatmap y-axis row labels. The column/s must
+            produce a unique index.
+        max_len : int, optional
+            Displaying large styled dataframes may cause ipython notebooks to
+            crash.
+        x_label : str, optional
+            This is the x-axis label that will be displayed on the heatmap.
+        y_label : str, optional
+            This is the y-axis label that will be displayed on the heatmap.
+        colorbar : bool, optional
+            If False, colorbar is not output.
+        col_width : int, optional
+            Plot width per column in pixels (px).
+        width : int, optional
+            Plot width in pixels (px), overrides col_width.
+        row_height : int, optional
+            Plot height per row in pixels (px).
+        height : int, optional
+            Plot height in pixels (px), overrides row_height.
+        text_auto : str, optional
+            Formatting for frequency values.
+        aspect : str, optional
+            Control the aspect ratio of the heatmap.
+        color_continuous_scale : str, optional
+            Color scale to use.
+        title : bool or str, optional
+            If True, attempt to use metadata from input dataset as a plot
+            title. Otherwise, use supplied value as a title.
+        **kwargs
+            Other parameters are passed through to px.imshow().
+
+        Returns
+        -------
+        fig : plotly.graph_objects.Figure
+
+        """
+        debug = self._log.debug
+
+        import plotly.express as px
+
+        debug("check len of input")
+        if len(df) > max_len:
+            raise ValueError(f"Input DataFrame is longer than {max_len}")
+
+        debug("handle title")
+        if title is True:
+            title = df.attrs.get("title", None)
+
+        debug("indexing")
+        if index is None:
+            index = list(df.index.names)
+        df = df.reset_index().copy()
+        if isinstance(index, list):
+            index_col = (
+                df[index]
+                .astype(str)
+                .apply(
+                    lambda row: ", ".join([o for o in row if o is not None]),
+                    axis="columns",
+                )
+            )
+        elif isinstance(index, str):
+            index_col = df[index].astype(str)
+        else:
+            raise TypeError("wrong type for index parameter, expected list or str")
+
+        debug("check that index is unique")
+        if not index_col.is_unique:
+            raise ValueError(f"{index} does not produce a unique index")
+
+        debug("drop and re-order columns")
+        frq_cols = [col for col in df.columns if col.startswith("frq_")]
+
+        debug("keep only freq cols")
+        heatmap_df = df[frq_cols].copy()
+
+        debug("set index")
+        heatmap_df.set_index(index_col, inplace=True)
+
+        debug("clean column names")
+        heatmap_df.columns = heatmap_df.columns.str.lstrip("frq_")
+
+        debug("deal with width and height")
+        if width is None:
+            width = 400 + col_width * len(heatmap_df.columns)
+            if colorbar:
+                width += 40
+        if height is None:
+            height = 200 + row_height * len(heatmap_df)
+            if title is not None:
+                height += 40
+
+        debug("plotly heatmap styling")
+        fig = px.imshow(
+            img=heatmap_df,
+            zmin=0,
+            zmax=1,
+            width=width,
+            height=height,
+            text_auto=text_auto,
+            aspect=aspect,
+            color_continuous_scale=color_continuous_scale,
+            title=title,
+            **kwargs,
+        )
+
+        fig.update_xaxes(side="bottom", tickangle=30)
+        if x_label is not None:
+            fig.update_xaxes(title=x_label)
+        if y_label is not None:
+            fig.update_yaxes(title=y_label)
+        fig.update_layout(
+            coloraxis_colorbar=dict(
+                title="Frequency",
+                tickvals=[0, 0.2, 0.4, 0.6, 0.8, 1.0],
+                ticktext=["0%", "20%", "40%", "60%", "80%", "100%"],
+            )
+        )
+        if not colorbar:
+            fig.update(layout_coloraxis_showscale=False)
+
+        return fig
+
+    def plot_frequencies_time_series(
+        self, ds, height=None, width=None, title=True, **kwargs
+    ):
+        """Create a time series plot of variant frequencies using plotly.
+
+        Parameters
+        ----------
+        ds : xarray.Dataset
+            A dataset of variant frequencies, such as returned by
+            `snp_allele_frequencies_advanced()`,
+            `aa_allele_frequencies_advanced()` or
+            `gene_cnv_frequencies_advanced()`.
+        height : int, optional
+            Height of plot in pixels (px).
+        width : int, optional
+            Width of plot in pixels (px).
+        title : bool or str, optional
+            If True, attempt to use metadata from input dataset as a plot
+            title. Otherwise, use supplied value as a title.
+        **kwargs
+            Passed through to `px.line()`.
+
+        Returns
+        -------
+        fig : plotly.graph_objects.Figure
+            A plotly figure containing line graphs. The resulting figure will
+            have one panel per cohort, grouped into columns by taxon, and
+            grouped into rows by area. Markers and lines show frequencies of
+            variants.
+
+        """
+        debug = self._log.debug
+
+        import plotly.express as px
+
+        debug("handle title")
+        if title is True:
+            title = ds.attrs.get("title", None)
+
+        debug("extract cohorts into a dataframe")
+        # FIXME: unresolved attribute reference 'startswith'
+        cohort_vars = [v for v in ds if v.startswith("cohort_")]
+        df_cohorts = ds[cohort_vars].to_dataframe()
+        df_cohorts.columns = [c.split("cohort_")[1] for c in df_cohorts.columns]
+
+        debug("extract variant labels")
+        variant_labels = ds["variant_label"].values
+
+        debug("build a long-form dataframe from the dataset")
+        dfs = []
+        for cohort_index, cohort in enumerate(df_cohorts.itertuples()):
+            ds_cohort = ds.isel(cohorts=cohort_index)
+            df = pd.DataFrame(
+                {
+                    "taxon": cohort.taxon,
+                    "area": cohort.area,
+                    "date": cohort.period_start,
+                    "period": str(
+                        cohort.period
+                    ),  # use string representation for hover label
+                    "sample_size": cohort.size,
+                    "variant": variant_labels,
+                    "count": ds_cohort["event_count"].values,
+                    "nobs": ds_cohort["event_nobs"].values,
+                    "frequency": ds_cohort["event_frequency"].values,
+                    "frequency_ci_low": ds_cohort["event_frequency_ci_low"].values,
+                    "frequency_ci_upp": ds_cohort["event_frequency_ci_upp"].values,
+                }
+            )
+            dfs.append(df)
+        df_events = pd.concat(dfs, axis=0).reset_index(drop=True)
+
+        debug("remove events with no observations")
+        df_events = df_events.query("nobs > 0")
+
+        debug("calculate error bars")
+        frq = df_events["frequency"]
+        frq_ci_low = df_events["frequency_ci_low"]
+        frq_ci_upp = df_events["frequency_ci_upp"]
+        df_events["frequency_error"] = frq_ci_upp - frq
+        df_events["frequency_error_minus"] = frq - frq_ci_low
+
+        debug("make a plot")
+        fig = px.line(
+            df_events,
+            facet_col="taxon",
+            facet_row="area",
+            x="date",
+            y="frequency",
+            error_y="frequency_error",
+            error_y_minus="frequency_error_minus",
+            color="variant",
+            markers=True,
+            hover_name="variant",
+            hover_data={
+                "frequency": ":.0%",
+                "period": True,
+                "area": True,
+                "taxon": True,
+                "sample_size": True,
+                "date": False,
+                "variant": False,
+            },
+            height=height,
+            width=width,
+            title=title,
+            labels={
+                "date": "Date",
+                "frequency": "Frequency",
+                "variant": "Variant",
+                "taxon": "Taxon",
+                "area": "Area",
+                "period": "Period",
+                "sample_size": "Sample size",
+            },
+            **kwargs,
+        )
+
+        debug("tidy plot")
+        fig.update_layout(yaxis_range=[-0.05, 1.05])
+
+        return fig
+
+    def plot_frequencies_map_markers(self, m, ds, variant, taxon, period, clear=True):
+        """Plot markers on a map showing variant frequencies for cohorts grouped
+        by area (space), period (time) and taxon.
+
+        Parameters
+        ----------
+        m : ipyleaflet.Map
+            The map on which to add the markers.
+        ds : xarray.Dataset
+            A dataset of variant frequencies, such as returned by
+            `snp_allele_frequencies_advanced()`,
+            `aa_allele_frequencies_advanced()` or
+            `gene_cnv_frequencies_advanced()`.
+        variant : int or str
+            Index or label of variant to plot.
+        taxon : str
+            Taxon to show markers for.
+        period : pd.Period
+            Time period to show markers for.
+        clear : bool, optional
+            If True, clear all layers (except the base layer) from the map
+            before adding new markers.
+
+        """
+        debug = self._log.debug
+
+        import ipyleaflet
+        import ipywidgets
+
+        debug("slice dataset to variant of interest")
+        if isinstance(variant, int):
+            ds_variant = ds.isel(variants=variant)
+            variant_label = ds["variant_label"].values[variant]
+        elif isinstance(variant, str):
+            ds_variant = ds.set_index(variants="variant_label").sel(variants=variant)
+            variant_label = variant
+        else:
+            raise TypeError(
+                f"Bad type for variant parameter; expected int or str, found {type(variant)}."
+            )
+
+        debug("convert to a dataframe for convenience")
+        df_markers = ds_variant[
+            [
+                "cohort_taxon",
+                "cohort_area",
+                "cohort_period",
+                "cohort_lat_mean",
+                "cohort_lon_mean",
+                "cohort_size",
+                "event_frequency",
+                "event_frequency_ci_low",
+                "event_frequency_ci_upp",
+            ]
+        ].to_dataframe()
+
+        debug("select data matching taxon and period parameters")
+        df_markers = df_markers.loc[
+            (
+                (df_markers["cohort_taxon"] == taxon)
+                & (df_markers["cohort_period"] == period)
+            )
+        ]
+
+        debug("clear existing layers in the map")
+        if clear:
+            for layer in m.layers[1:]:
+                m.remove_layer(layer)
+
+        debug("add markers")
+        for x in df_markers.itertuples():
+            marker = ipyleaflet.CircleMarker()
+            marker.location = (x.cohort_lat_mean, x.cohort_lon_mean)
+            marker.radius = 20
+            marker.color = "black"
+            marker.weight = 1
+            marker.fill_color = "red"
+            marker.fill_opacity = x.event_frequency
+            popup_html = f"""
+                <strong>{variant_label}</strong> <br/>
+                Taxon: {x.cohort_taxon} <br/>
+                Area: {x.cohort_area} <br/>
+                Period: {x.cohort_period} <br/>
+                Sample size: {x.cohort_size} <br/>
+                Frequency: {x.event_frequency:.0%}
+                (95% CI: {x.event_frequency_ci_low:.0%} - {x.event_frequency_ci_upp:.0%})
+            """
+            marker.popup = ipyleaflet.Popup(
+                child=ipywidgets.HTML(popup_html),
+            )
+            m.add_layer(marker)
+
+    def plot_frequencies_interactive_map(
+        self,
+        ds,
+        center=(-2, 20),
+        zoom=3,
+        title=True,
+        epilogue=True,
+    ):
+        """Create an interactive map with markers showing variant frequencies or
+        cohorts grouped by area (space), period (time) and taxon.
+
+        Parameters
+        ----------
+        ds : xarray.Dataset
+            A dataset of variant frequencies, such as returned by
+            `snp_allele_frequencies_advanced()`,
+            `aa_allele_frequencies_advanced()` or
+            `gene_cnv_frequencies_advanced()`.
+        center : tuple of int, optional
+            Location to center the map.
+        zoom : int, optional
+            Initial zoom level.
+        title : bool or str, optional
+            If True, attempt to use metadata from input dataset as a plot
+            title. Otherwise, use supplied value as a title.
+        epilogue : bool or str, optional
+            Additional text to display below the map.
+
+        Returns
+        -------
+        out : ipywidgets.Widget
+            An interactive map with widgets for selecting which variant, taxon
+            and time period to display.
+
+        """
+        debug = self._log.debug
+
+        import ipyleaflet
+        import ipywidgets
+
+        debug("handle title")
+        if title is True:
+            title = ds.attrs.get("title", None)
+
+        debug("create a map")
+        freq_map = ipyleaflet.Map(center=center, zoom=zoom)
+
+        debug("set up interactive controls")
+        variants = ds["variant_label"].values
+        taxa = np.unique(ds["cohort_taxon"].values)
+        periods = np.unique(ds["cohort_period"].values)
+        controls = ipywidgets.interactive(
+            self.plot_frequencies_map_markers,
+            m=ipywidgets.fixed(freq_map),
+            ds=ipywidgets.fixed(ds),
+            variant=ipywidgets.Dropdown(options=variants, description="Variant: "),
+            taxon=ipywidgets.Dropdown(options=taxa, description="Taxon: "),
+            period=ipywidgets.Dropdown(options=periods, description="Period: "),
+            clear=ipywidgets.fixed(True),
+        )
+
+        debug("lay out widgets")
+        components = []
+        if title is not None:
+            components.append(ipywidgets.HTML(value=f"<h3>{title}</h3>"))
+        components.append(controls)
+        components.append(freq_map)
+        if epilogue is True:
+            epilogue = """
+                Variant frequencies are shown as coloured markers. Opacity of color
+                denotes frequency. Click on a marker for more information.
+            """
+        if epilogue:
+            components.append(ipywidgets.HTML(value=f"{epilogue}"))
+
+        out = ipywidgets.VBox(components)
+
+        return out
+
+    def plot_pca_coords(
+        self,
+        data,
+        x="PC1",
+        y="PC2",
+        color=None,
+        symbol=None,
+        jitter_frac=0.02,
+        random_seed=42,
+        width=900,
+        height=600,
+        marker_size=10,
+        **kwargs,
+    ):
+        """Plot sample coordinates from a principal components analysis (PCA)
+        as a plotly scatter plot.
+
+        Parameters
+        ----------
+        data : pandas.DataFrame
+            A dataframe of sample metadata, with columns "PC1", "PC2", "PC3",
+            etc., added.
+        x : str, optional
+            Name of principal component to plot on the X axis.
+        y : str, optional
+            Name of principal component to plot on the Y axis.
+        color : str, optional
+            Name of column in the input dataframe to use to color the markers.
+        symbol : str, optional
+            Name of column in the input dataframe to use to choose marker symbols.
+        jitter_frac : float, optional
+            Randomly jitter points by this fraction of their range.
+        random_seed : int, optional
+            Random seed for jitter.
+        width : int, optional
+            Plot width in pixels (px).
+        height : int, optional
+            Plot height in pixels (px).
+        marker_size : int, optional
+            Marker size.
+
+        Returns
+        -------
+        fig : Figure
+            A plotly figure.
+
+        """
+        debug = self._log.debug
+
+        import plotly.express as px
+
+        debug(
+            "set up data - copy and shuffle so that we don't get systematic over-plotting"
+        )
+        # TODO does the shuffling actually work?
+        data = (
+            data.copy().sample(frac=1, random_state=random_seed).reset_index(drop=True)
+        )
+
+        debug(
+            "apply jitter if desired - helps spread out points when tightly clustered"
+        )
+        if jitter_frac:
+            np.random.seed(random_seed)
+            data[x] = jitter(data[x], jitter_frac)
+            data[y] = jitter(data[y], jitter_frac)
+
+        debug("convenience variables")
+        data["country_location"] = data["country"] + " - " + data["location"]
+
+        debug("set up plotting options")
+        hover_data = [
+            "partner_sample_id",
+            "sample_set",
+            "taxon",
+            "country",
+            "admin1_iso",
+            "admin1_name",
+            "admin2_name",
+            "location",
+            "year",
+            "month",
+        ]
+        plot_kwargs = dict(
+            width=width,
+            height=height,
+            color=color,
+            symbol=symbol,
+            template="simple_white",
+            hover_name="sample_id",
+            hover_data=hover_data,
+            opacity=0.9,
+            render_mode="svg",
+        )
+
+        debug("special handling for taxon color")
+        if color == "taxon":
+            self._setup_taxon_colors(plot_kwargs)
+
+        debug("apply any user overrides")
+        plot_kwargs.update(kwargs)
+
+        debug("2D scatter plot")
+        fig = px.scatter(data, x=x, y=y, **plot_kwargs)
+
+        debug("tidy up")
+        fig.update_layout(
+            legend=dict(itemsizing="constant"),
+        )
+        fig.update_traces(marker={"size": marker_size})
+
+        return fig
+
+    def plot_pca_coords_3d(
+        self,
+        data,
+        x="PC1",
+        y="PC2",
+        z="PC3",
+        color=None,
+        symbol=None,
+        jitter_frac=0.02,
+        random_seed=42,
+        width=900,
+        height=600,
+        marker_size=5,
+        **kwargs,
+    ):
+        """Plot sample coordinates from a principal components analysis (PCA)
+        as a plotly 3D scatter plot.
+
+        Parameters
+        ----------
+        data : pandas.DataFrame
+            A dataframe of sample metadata, with columns "PC1", "PC2", "PC3",
+            etc., added.
+        x : str, optional
+            Name of principal component to plot on the X axis.
+        y : str, optional
+            Name of principal component to plot on the Y axis.
+        z : str, optional
+            Name of principal component to plot on the Z axis.
+        color : str, optional
+            Name of column in the input dataframe to use to color the markers.
+        symbol : str, optional
+            Name of column in the input dataframe to use to choose marker symbols.
+        jitter_frac : float, optional
+            Randomly jitter points by this fraction of their range.
+        random_seed : int, optional
+            Random seed for jitter.
+        width : int, optional
+            Plot width in pixels (px).
+        height : int, optional
+            Plot height in pixels (px).
+        marker_size : int, optional
+            Marker size.
+
+        Returns
+        -------
+        fig : Figure
+            A plotly figure.
+
+        """
+        debug = self._log.debug
+
+        import plotly.express as px
+
+        debug(
+            "set up data - copy and shuffle so that we don't get systematic over-plotting"
+        )
+        # TODO does this actually work?
+        data = (
+            data.copy().sample(frac=1, random_state=random_seed).reset_index(drop=True)
+        )
+
+        debug(
+            "apply jitter if desired - helps spread out points when tightly clustered"
+        )
+        if jitter_frac:
+            np.random.seed(random_seed)
+            data[x] = jitter(data[x], jitter_frac)
+            data[y] = jitter(data[y], jitter_frac)
+            data[z] = jitter(data[z], jitter_frac)
+
+        debug("convenience variables")
+        data["country_location"] = data["country"] + " - " + data["location"]
+
+        debug("set up plotting options")
+        hover_data = [
+            "partner_sample_id",
+            "sample_set",
+            "taxon",
+            "country",
+            "admin1_iso",
+            "admin1_name",
+            "admin2_name",
+            "location",
+            "year",
+            "month",
+        ]
+        plot_kwargs = dict(
+            width=width,
+            height=height,
+            hover_name="sample_id",
+            hover_data=hover_data,
+            color=color,
+            symbol=symbol,
+        )
+
+        debug("special handling for taxon color")
+        if color == "taxon":
+            self._setup_taxon_colors(plot_kwargs)
+
+        debug("apply any user overrides")
+        plot_kwargs.update(kwargs)
+
+        debug("3D scatter plot")
+        fig = px.scatter_3d(data, x=x, y=y, z=z, **plot_kwargs)
+
+        debug("tidy up")
+        fig.update_layout(
+            scene=dict(aspectmode="cube"),
+            legend=dict(itemsizing="constant"),
+        )
+        fig.update_traces(marker={"size": marker_size})
+
+        return fig
+
+    def plot_diversity_stats(
+        self,
+        df_stats,
+        color=None,
+        bar_plot_height=450,
+        bar_width=30,
+        scatter_plot_height=500,
+        scatter_plot_width=500,
+        template="plotly_white",
+        plot_kwargs=None,
+    ):
+        """Plot diversity statistics.
+
+        Parameters
+        ----------
+        df_stats : pandas.DataFrame
+            Output from diversity_stats().
+        color : str, optional
+            Column to color by.
+        bar_plot_height : int, optional
+            Height of bar plots in pixels (px).
+        bar_width : int, optional
+            Width per bar in pixels (px).
+        scatter_plot_height : int, optional
+            Height of scatter plot in pixels (px).
+        scatter_plot_width : int, optional
+            Width of scatter plot in pixels (px).
+        template : str, optional
+            Plotly template.
+        plot_kwargs : dict, optional
+            Extra plotting parameters
+
+        """
+        debug = self._log.debug
+        import plotly.express as px
+
+        debug("set up common plotting parameters")
+        if plot_kwargs is None:
+            plot_kwargs = dict()
+        default_plot_kwargs = dict(
+            hover_name="cohort",
+            hover_data=[
+                "taxon",
+                "country",
+                "admin1_iso",
+                "admin1_name",
+                "admin2_name",
+                "longitude",
+                "latitude",
+                "year",
+                "month",
+            ],
+            labels={
+                "theta_pi_estimate": r"$\widehat{\theta}_{\pi}$",
+                "theta_w_estimate": r"$\widehat{\theta}_{w}$",
+                "tajima_d_estimate": r"$D$",
+                "cohort": "Cohort",
+                "taxon": "Taxon",
+                "country": "Country",
+            },
+        )
+        if color == "taxon":
+            self._setup_taxon_colors(plot_kwargs=default_plot_kwargs)
+        default_plot_kwargs.update(plot_kwargs)
+        plot_kwargs = default_plot_kwargs
+        bar_plot_width = 300 + bar_width * len(df_stats)
+
+        debug("nucleotide diversity bar plot")
+        fig = px.bar(
+            data_frame=df_stats,
+            x="cohort",
+            y="theta_pi_estimate",
+            error_y="theta_pi_ci_err",
+            title="Nucleotide diversity",
+            color=color,
+            height=bar_plot_height,
+            width=bar_plot_width,
+            template=template,
+            **plot_kwargs,
+        )
+        fig.show()
+
+        debug("Watterson's estimator bar plot")
+        fig = px.bar(
+            data_frame=df_stats,
+            x="cohort",
+            y="theta_w_estimate",
+            error_y="theta_w_ci_err",
+            title="Watterson's estimator",
+            color=color,
+            height=bar_plot_height,
+            width=bar_plot_width,
+            template=template,
+            **plot_kwargs,
+        )
+        fig.show()
+
+        debug("Tajima's D bar plot")
+        fig = px.bar(
+            data_frame=df_stats,
+            x="cohort",
+            y="tajima_d_estimate",
+            error_y="tajima_d_ci_err",
+            title="Tajima's D",
+            color=color,
+            height=bar_plot_height,
+            width=bar_plot_width,
+            template=template,
+            **plot_kwargs,
+        )
+        fig.show()
+
+        debug("scatter plot comparing diversity estimators")
+        fig = px.scatter(
+            data_frame=df_stats,
+            x="theta_pi_estimate",
+            y="theta_w_estimate",
+            error_x="theta_pi_ci_err",
+            error_y="theta_w_ci_err",
+            title="Diversity estimators",
+            color=color,
+            width=scatter_plot_width,
+            height=scatter_plot_height,
+            template=template,
+            **plot_kwargs,
+        )
+        fig.show()
+
+    def plot_samples_interactive_map(
+        self,
+        sample_sets=None,
+        sample_query=None,
+        basemap=None,
+        center=(-2, 20),
+        zoom=3,
+        min_samples=1,
+    ):
+        """Plot an interactive map showing sampling locations using ipyleaflet.
+
+        Parameters
+        ----------
+        sample_sets : str or list of str, optional
+            Can be a sample set identifier or a list of
+            sample set identifiers or a
+            release identifier or a list of release identifiers.
+        sample_query : str, optional
+            A pandas query string which will be evaluated against the sample
+            metadata e.g., "taxon == 'coluzzii' and country == 'Burkina Faso'".
+        basemap : dict
+            Basemap description coming from ipyleaflet.basemaps.
+        center : tuple of int, optional
+            Location to center the map.
+        zoom : int, optional
+            Initial zoom level.
+        min_samples : int, optional
+            Minimum number of samples required to show a marker for a given
+            location.
+
+        Returns
+        -------
+        samples_map : ipyleaflet.Map
+            Ipyleaflet map widget.
+
+        """
+        debug = self._log.debug
+
+        import ipyleaflet
+
+        debug("load sample metadata")
+        df_samples = self.sample_metadata(
+            sample_sets=sample_sets, sample_query=sample_query
+        )
+
+        debug("compute locations")
+        pivot_location_taxon = df_samples.pivot_table(
+            index=[
+                "country",
+                "admin1_iso",
+                "admin1_name",
+                "admin2_name",
+                "location",
+                "latitude",
+                "longitude",
+            ],
+            columns=["taxon"],
+            values="sample_id",
+            aggfunc="count",
+            fill_value=0,
+        )
+
+        debug("create a map")
+        if basemap is None:
+            # FIXME: cannot find reference 'Esri'
+            basemap = ipyleaflet.basemaps.Esri.WorldImagery
+        samples_map = ipyleaflet.Map(
+            center=center,
+            zoom=zoom,
+            basemap=basemap,
+        )
+        samples_map.add_control(ipyleaflet.ScaleControl(position="bottomleft"))
+        # make the map a bit taller than the default
+        samples_map.layout.height = "500px"
+
+        debug("add markers")
+        taxa = df_samples["taxon"].dropna().sort_values().unique()
+        for _, row in pivot_location_taxon.reset_index().iterrows():
+            title = (
+                f"Location: {row.location} ({row.latitude:.3f}, {row.longitude:.3f})"
+            )
+            title += f"\nAdmin level 2: {row.admin2_name}"
+            title += f"\nAdmin level 1: {row.admin1_name} ({row.admin1_iso})"
+            title += f"\nCountry: {row.country}"
+            title += "\nNo. specimens: "
+            all_n = 0
+            for taxon in taxa:
+                n = row[taxon]
+                all_n += n
+                if n > 0:
+                    title += f"{n} {taxon}; "
+            if all_n >= min_samples:
+                marker = ipyleaflet.Marker(
+                    location=(row.latitude, row.longitude),
+                    draggable=False,
+                    title=title,
+                )
+                samples_map.add_layer(marker)
+
+        return samples_map
+
+    def plot_fst_gwss_track(
+        self,
+        contig,
+        window_size,
+        cohort1_query,
+        cohort2_query,
+        sample_sets=None,
+        site_mask="default",
+        cohort_size=30,
+        random_seed=42,
+        title=None,
+        width=DEFAULT_GENOME_PLOT_WIDTH,
+        height=200,
+        show=True,
+        x_range=None,
+    ):
+        """Run and plot a Fst genome-wide scan to investigate genetic
+            differentiation between two cohorts.
+        Parameters
+        ----------
+        contig: str
+            Contig name (e.g., "X")
+        window_size : int
+            The size of windows used to calculate h12 over.
+        cohort1_query : str
+            A pandas query string which will be evaluated against the sample
+            metadata e.g., "taxon == 'coluzzii' and country == 'Burkina Faso'".
+        cohort2_query : str
+            A pandas query string which will be evaluated against the sample
+            metadata e.g., "taxon == 'coluzzii' and country == 'Burkina Faso'".
+        sample_sets : str or list of str, optional
+            Can be a sample set identifier or a list of
+            sample set identifiers or a
+            release identifier or a list of release identifiers.
+        site_mask : str, optional
+            Site filters mask to apply.
+        cohort_size : int, optional
+            If provided, randomly down-sample to the given cohort size.
+        random_seed : int, optional
+            Random seed used for down-sampling.
+        title : str, optional
+            If provided, title string is used to label plot.
+        width : int, optional
+            Plot width in pixels (px).
+        height : int. optional
+            Plot height in pixels (px).
+        show : bool, optional
+            If True, show the plot.
+        x_range : bokeh.models.Range1d, optional
+            X axis range (for linking to other tracks).
+
+        Returns
+        -------
+        fig : figure
+            A plot showing windowed Fst statistic across chosen contig.
+        """
+
+        import bokeh.models as bkmod
+        import bokeh.plotting as bkplt
+
+        # compute Fst
+        x, fst = self.fst_gwss(
+            contig=contig,
+            window_size=window_size,
+            cohort_size=cohort_size,
+            cohort1_query=cohort1_query,
+            cohort2_query=cohort2_query,
+            sample_sets=sample_sets,
+            site_mask=site_mask,
+            random_seed=random_seed,
+        )
+
+        # determine X axis range
+        x_min = x[0]
+        x_max = x[-1]
+        if x_range is None:
+            x_range = bkmod.Range1d(x_min, x_max, bounds="auto")
+
+        # create a figure
+        xwheel_zoom = bkmod.WheelZoomTool(dimensions="width", maintain_focus=False)
+        if title is None:
+            title = f"Cohort 1: {cohort1_query}\nCohort 2: {cohort2_query}"
+        fig = bkplt.figure(
+            title=title,
+            tools=["xpan", "xzoom_in", "xzoom_out", xwheel_zoom, "reset"],
+            active_scroll=xwheel_zoom,
+            active_drag="xpan",
+            plot_width=width,
+            plot_height=height,
+            toolbar_location="above",
+            x_range=x_range,
+            y_range=(0, 1),
+        )
+
+        # plot Fst
+        fig.circle(
+            x=x,
+            y=fst,
+            size=3,
+            line_width=0.5,
+            line_color="black",
+            fill_color=None,
+        )
+
+        # tidy up the plot
+        fig.yaxis.axis_label = "Fst"
+        fig.yaxis.ticker = [0, 1]
+        self._bokeh_style_genome_xaxis(fig, contig)
+
+        if show:
+            bkplt.show(fig)
+
+        return fig
+
+    def plot_fst_gwss(
+        self,
+        contig,
+        window_size,
+        cohort1_query,
+        cohort2_query,
+        sample_sets=None,
+        site_mask="default",
+        cohort_size=30,
+        random_seed=42,
+        title=None,
+        width=DEFAULT_GENOME_PLOT_WIDTH,
+        track_height=190,
+        genes_height=DEFAULT_GENES_TRACK_HEIGHT,
+    ):
+        """Run and plot a Fst genome-wide scan to investigate genetic
+        differentiation between two cohorts.
+
+        Parameters
+        ----------
+        contig: str
+            Contig name (e.g., "X")
+        window_size : int
+            The size of windows used to calculate h12 over.
+        cohort1_query : str
+            A pandas query string which will be evaluated against the sample
+            metadata e.g., "taxon == 'coluzzii' and country == 'Burkina Faso'".
+        cohort2_query : str
+            A pandas query string which will be evaluated against the sample
+            metadata e.g., "taxon == 'coluzzii' and country == 'Burkina Faso'".
+        sample_sets : str or list of str, optional
+            Can be a sample set identifier or a list of
+            sample set identifiers or a
+            release identifier or a list of release identifiers.
+        site_mask : str, optional
+            Site filters mask to apply.
+        cohort_size : int, optional
+            If provided, randomly down-sample to the given cohort size.
+        random_seed : int, optional
+            Random seed used for down-sampling.
+        title : str, optional
+            If provided, title string is used to label plot.
+        width : int, optional
+            Plot width in pixels (px).
+        track_height : int. optional
+            GWSS track height in pixels (px).
+        genes_height : int. optional
+            Gene track height in pixels (px).
+
+        Returns
+        -------
+        fig : figure
+            A plot showing windowed Fst statistic with gene track on x-axis.
+        """
+
+        import bokeh.layouts as bklay
+        import bokeh.plotting as bkplt
+
+        # gwss track
+        fig1 = self.plot_fst_gwss_track(
+            contig=contig,
+            window_size=window_size,
+            cohort1_query=cohort1_query,
+            cohort2_query=cohort2_query,
+            sample_sets=sample_sets,
+            site_mask=site_mask,
+            cohort_size=cohort_size,
+            random_seed=random_seed,
+            title=title,
+            width=width,
+            height=track_height,
+            show=False,
+        )
+
+        fig1.xaxis.visible = False
+
+        # plot genes
+        fig2 = self.plot_genes(
+            region=contig,
+            width=width,
+            height=genes_height,
+            x_range=fig1.x_range,
+            show=False,
+        )
+
+        # combine plots into a single figure
+        fig = bklay.gridplot(
+            [fig1, fig2], ncols=1, toolbar_location="above", merge_tools=True
+        )
+
+        bkplt.show(fig)
